backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.25.1
redis==5.0.1
xata==1.0.0
python-dotenv==1.0.0
email-validator==2.1.0 